requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.4
django-cors-headers==3.5.0
django-on-heroku==1.1.2
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==3.1.1
psycopg2-binary==2.8.6
pytz==2020.4
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.2
whitenoise==5.2.0
zipp==3.4.0
